Add restart game and separate index.js to new files
