+++ index.html
@@ -6,7 +6,8 @@
         <meta name="viewport" content="width=device-width, initial-scale=1.0" />
         <link rel="stylesheet" href="style.css" />
         <title>Document</title>
-        <script defer src="index.js"></script>
+        <script type="module" defer src="index.js"></script>
+
         <link rel="preconnect" href="https://fonts.googleapis.com" />
         <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
         <link
@@ -36,5 +37,8 @@
             </div>
         </div>
         <p id="round">Tura:X</p>
+        <p id="playAgain" class="hidden">
+            Kliknij na mnie aby zagrać ponownie!
+        </p>
     </body>
 </html>
