+++ style.css
@@ -7,8 +7,13 @@
     }
 
     .blank {
-        width: 200px;
-        height: 200px;
+        width: 150px;
+        height: 150px;
+    }
+
+    #round {
+        font-size: 20px;
+        margin: 3px;
     }
 }
 
@@ -22,8 +27,8 @@
     }
 
     .blank {
-        width: 150px;
-        height: 150px;
+        width: 100px;
+        height: 100px;
     }
 }
 
@@ -103,4 +108,18 @@ html {
 
 .off {
     cursor: auto;
+}
+
+#playAgain {
+    font-size: 18px;
+    cursor: pointer;
+    transition: color 0.2s linear;
+}
+
+#playAgain:hover {
+    color: #00aa00;
+}
+
+.hidden {
+    display: none;
 }
